wrapped-api/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-dotenv
sqlite-utils
pydantic
aiolimiter
boto3